Passport y JWT

+++ controllers/productosControllers.js
@@ -33,9 +33,12 @@ const productosControllers = {
     },
     cargarUnProducto: (req,res)=>{
         const {name, precio, description, stock} = req.body
-        
-        new Producto({name, precio, description, stock}).save()
-        .then((respuesta) => res.json({respuesta}))
+        if (req.user) {
+            new Producto({name, precio, description, stock}).save()
+            .then((respuesta) => res.json({respuesta}))
+        }else{
+            res.json({success: true, error:'El usuario no es autentico, prueba loguearte de nuevo.'})
+        }
         
 
         // res.json({respuesta: productos})
@@ -43,15 +46,20 @@ const productosControllers = {
     borrarUnProducto: async(req,res)=>{
         const id = req.params.id
         let productos
-        try{
-            await Producto.findOneAndDelete({_id:id})
-            productos = await Producto.find()
+        console.log(req.user)
+        if (req.user) {
+            try{
+                await Producto.findOneAndDelete({_id:id})
+                productos = await Producto.find()
 
-        }catch(error){
-            console.log(error)
+            }catch(error){
+                console.log(error)
+            }
+            res.json({respuesta: productos,success:true})
+        }else{
+            res.json({success: true, error:'El usuario no es autentico'})
         }
 
-        res.json({respuesta: productos,success:true})
     },
     modificarProducto: async(req,res)=>{
         let id = req.params.id
